refactor(pinsetstore): split IO operations from types.rs to codec.rs and split types and codec tests
- split generic_array from types.rs into a helper directory
