+++ crates/backend/src/pinset/mod.rs
@@ -1,6 +1,5 @@
+mod codec;
+mod helpers;
 pub mod types;
-
-#[cfg(test)]
-mod tests;
 
 // add public API surface
